refactor(#352): postSrcap -> postScrap 패키지 오타 수정
