+++ src/main/java/com/cotato/kampus/domain/post/application/PostInteractionService.java
@@ -18,9 +18,9 @@ import com.cotato.kampus.domain.post.implement.postLike.PostLikeAppender;
 import com.cotato.kampus.domain.post.implement.postLike.PostLikeDeleter;
 import com.cotato.kampus.domain.post.implement.postLike.PostLikeFinder;
 import com.cotato.kampus.domain.post.implement.postLike.PostLikeValidator;
-import com.cotato.kampus.domain.post.implement.postSrcap.PostScrapAppender;
-import com.cotato.kampus.domain.post.implement.postSrcap.PostScrapDeleter;
-import com.cotato.kampus.domain.post.implement.postSrcap.PostScrapFinder;
+import com.cotato.kampus.domain.post.implement.postScrap.PostScrapAppender;
+import com.cotato.kampus.domain.post.implement.postScrap.PostScrapDeleter;
+import com.cotato.kampus.domain.post.implement.postScrap.PostScrapFinder;
 import com.cotato.kampus.domain.post.implement.trendingPost.TrendingPostAppender;
 import com.cotato.kampus.domain.post.implement.trendingPost.TrendingPostDeleter;
 import com.cotato.kampus.domain.post.implement.trendingPost.TrendingPostFinder;
