+++ src/main/java/com/cotato/kampus/domain/post/application/PostService.java
@@ -32,8 +32,8 @@ import com.cotato.kampus.domain.post.implement.postLike.PostLikeFinder;
 import com.cotato.kampus.domain.post.implement.postImage.PostPhotoAppender;
 import com.cotato.kampus.domain.post.implement.postImage.PostPhotoDeleter;
 import com.cotato.kampus.domain.post.implement.postImage.PostPhotoFinder;
-import com.cotato.kampus.domain.post.implement.postSrcap.PostScrapDeleter;
-import com.cotato.kampus.domain.post.implement.postSrcap.PostScrapFinder;
+import com.cotato.kampus.domain.post.implement.postScrap.PostScrapDeleter;
+import com.cotato.kampus.domain.post.implement.postScrap.PostScrapFinder;
 import com.cotato.kampus.domain.post.domain.PostDetails;
 import com.cotato.kampus.domain.post.domain.PostThumbnailWithBoardName;
 import com.cotato.kampus.domain.post.implement.trendingPost.TrendingPostDeleter;
