+++ src/main/java/com/cotato/kampus/domain/post/implement/postScrap/PostScrapDeleter.java
@@ -1,6 +1,4 @@
-package com.cotato.kampus.domain.post.implement.postSrcap;
-
-import java.util.List;
+package com.cotato.kampus.domain.post.implement.postScrap;
 
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
