+++ src/test/java/com/cotato/kampus/domain/post/application/PostInteractionServiceTest.java
@@ -7,9 +7,9 @@ import com.cotato.kampus.domain.post.domain.TestPostHelper;
 import com.cotato.kampus.domain.post.implement.post.PostFinder;
 import com.cotato.kampus.domain.post.implement.post.PostUpdater;
 import com.cotato.kampus.domain.post.implement.post.PostValidator;
-import com.cotato.kampus.domain.post.implement.postSrcap.PostScrapAppender;
-import com.cotato.kampus.domain.post.implement.postSrcap.PostScrapDeleter;
-import com.cotato.kampus.domain.post.implement.postSrcap.PostScrapFinder;
+import com.cotato.kampus.domain.post.implement.postScrap.PostScrapAppender;
+import com.cotato.kampus.domain.post.implement.postScrap.PostScrapDeleter;
+import com.cotato.kampus.domain.post.implement.postScrap.PostScrapFinder;
 import com.cotato.kampus.global.error.ErrorCode;
 import com.cotato.kampus.global.error.exception.AppException;
 import org.junit.jupiter.api.DisplayName;
